Finally, text under the bottom of the page!
This was accomplished using a div for the main section of the page with height
set to 100%, forcing the other section beneath the page.

Did I ever tell you how much fun it is to develop on a Mac using brackets on a
laptop monitor and the live preview page on the other @markacastle123?